+++ test/cc/unit/envoy_config_test.cc
@@ -80,19 +80,6 @@ TEST(TestConfig, ConfigIsValid) {
 #endif
 }
 
-#if !defined(__APPLE__)
-TEST(TestConfig, SetUseDnsCAresResolver) {
-  EngineBuilder engine_builder;
-  engine_builder.useDnsSystemResolver(false);
-  std::string config_str = engine_builder.generateConfigStr();
-  envoy::config::bootstrap::v3::Bootstrap bootstrap;
-  TestUtility::loadFromYaml(absl::StrCat(config_header, config_str), bootstrap);
-
-  ASSERT_THAT(bootstrap.DebugString(), HasSubstr("envoy.network.dns_resolver.cares"));
-  ASSERT_THAT(bootstrap.DebugString(), Not(HasSubstr("envoy.network.dns_resolver.getaddrinfo")));
-}
-#endif
-
 TEST(TestConfig, SetGzip) {
   EngineBuilder engine_builder;
 
@@ -376,7 +363,49 @@ private:
   mutable int count_ = 0;
 };
 
-TEST(TestConfig, StringAccessors) {
+TEST(TestConfig, AddNativeFilter) {
+  EngineBuilder engine_builder;
+
+  std::string filter_name = "envoy.filters.http.buffer";
+  std::string filter_config =
+      "{\"@type\":\"type.googleapis.com/envoy.extensions.filters.http.buffer.v3.Buffer\","
+      "\"max_request_bytes\":5242880}";
+
+  std::string config_str = engine_builder.generateConfigStr();
+  ASSERT_THAT(config_str, Not(HasSubstr("- name: " + filter_name)));
+  ASSERT_THAT(config_str, Not(HasSubstr("  typed_config: " + filter_config)));
+  envoy::config::bootstrap::v3::Bootstrap bootstrap;
+  TestUtility::loadFromYaml(absl::StrCat(config_header, config_str), bootstrap);
+
+  engine_builder.addNativeFilter(filter_name, filter_config);
+
+  config_str = engine_builder.generateConfigStr();
+  ASSERT_THAT(config_str, HasSubstr("- name: " + filter_name));
+  ASSERT_THAT(config_str, HasSubstr("  typed_config: " + filter_config));
+  TestUtility::loadFromYaml(absl::StrCat(config_header, config_str), bootstrap);
+}
+
+TEST(TestConfig, AddPlatformFilter) {
+  EngineBuilder engine_builder;
+
+  std::string filter_name = "test_platform_filter";
+
+  std::string config_str = engine_builder.generateConfigStr();
+  ASSERT_THAT(config_str, Not(HasSubstr("http.platform_bridge.PlatformBridge")));
+  ASSERT_THAT(config_str, Not(HasSubstr("platform_filter_name: " + filter_name)));
+  envoy::config::bootstrap::v3::Bootstrap bootstrap;
+  TestUtility::loadFromYaml(absl::StrCat(config_header, config_str), bootstrap);
+
+  engine_builder.addPlatformFilter(filter_name);
+
+  config_str = engine_builder.generateConfigStr();
+  ASSERT_THAT(config_str, HasSubstr("http.platform_bridge.PlatformBridge"));
+  ASSERT_THAT(config_str, HasSubstr("platform_filter_name: " + filter_name));
+  TestUtility::loadFromYaml(absl::StrCat(config_header, config_str), bootstrap);
+}
+
+// TODO(RyanTheOptimist): This test seems to be flaky. #2641
+TEST(TestConfig, DISABLED_StringAccessors) {
   std::string name("accessor_name");
   EngineBuilder engine_builder;
   std::string data_string = "envoy string";